testing test for service

+++ backend/app/schemas/item_schema.py
@@ -4,19 +4,19 @@ from typing import Optional
 
 class ItemCreate(BaseModel):
     name: str
-    price: int
+    price: float
     barcode: str
 
 
 class ItemUpdate(BaseModel):
     name: Optional[str] = None
-    price: Optional[int] = None
+    price: Optional[float] = None
     barcode: Optional[str] = None
 
 
 class ItemOut(BaseModel):
     id: str
     name: str
-    price: int
+    price: float
     barcode: str
